requirements.txt
Flask
requests
python-dotenv
markupsafe
gunicorn